+++ CHANGES.md
@@ -4,6 +4,7 @@ I will try to keep track of changes between releases here
 
 ## Upcomming release, available in devel branch
 
+- fixed root password will expire in 0 days warning
 - added exit codes in common_AOK/usr_local_bin/network-check.sh
 - removed obsoleted common_AOK/usr_local_bin/battery-charge
 - aok - improved option handling when chrooted